web_cli/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
motor>=3.3.0
pymongo[srv]>=4.5.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.4.0
slowapi>=0.2.8
